Revert "Remove unnecessary runtime stream logging (#369)"
This reverts commit a406c8260794cd33910301a9c7e6d42ee06b0d9c.
