+++ packages/server/src/synchronizer/reporter.ts
@@ -2,13 +2,19 @@ import {through} from './streams'
 import File from 'vinyl'
 import {log} from '../log'
 import chalk from 'chalk'
-import {Event, FILE_WRITTEN, INIT, ERROR_THROWN, READY} from './events'
+
+export type Event<T> = {type: string; payload: T}
+
+export const INIT = 'INIT'
+export const FILE_WRITTEN = 'FILE_WRITTEN'
+export const ERROR_THROWN = 'ERROR_THROWN'
+export const READY = 'READY'
 
 /**
- * Display is a stream that converts build status events and prepares them for the console.
+ * Reporter is a stream that converts build status events and prepares them for the console.
  * A good way to think about this is as the root of the "view" component of the application.
  */
-export default function createDisplay() {
+export default function createReporter() {
   let lastEvent: Event<any> = {type: INIT, payload: null}
 
   let spinner = log.spinner('Preparing for launch').start()
